Splits writting process in three
 - write_status_line
 - write_headers
 - write

write_status_line and write_headers are both signals so that callbacks can be
connected after and before those important events. Also, they are called
automatically on first write and first write_headers.

No need to lock write operations: we do asynchronous I/O, not concurrent
writting.

Cookie implementation for Request and Response are virtuals. They should be
overwritten if the server provide a better implementation than looking up the
headers.

Uses properties for Request and Response headers.

Request provide a Soup.HTTPVersion property. The protocol is already accessible
through Request.uri.scheme.

FastCGI status is stored in Response headers.

Request instance in Response is a simple property so that it cannot be set
after the constructor.

Request and Response headers does not have to be abstract as we systematically
end up wrapping an instance of Soup.MessageHeaders whatever the implementation.