Uses debug instead of message to avoid undesired I/O.
Improves the throughput on all implementations by avoiding write
operations on the standard output by replacing calls to message for
calls to debug.

info is only available from glib >=2.40, so there is no way to produce
messages that does not get written to the stdout and remain compatible
with >=2.32.

Debug messages can be printed by setting the G_MESSAGES_DEBUG
environment variable to an appropriate value.

Removes the fastcgi.debug option from the lighttpd.conf to improve
performance in the same manner.